cant delete empty room

+++ www/models/Room.php
@@ -14,6 +14,7 @@ class Room
     }
 
     public static function createRoom($name, $homeId) {
+    
     $db=dbconnect();
     $stmt = $db->prepare("INSERT INTO rooms (name, id_home) VALUES (?, ?)");
 
@@ -26,10 +27,17 @@ class Room
     }
 
     public function deleteSelf() {
+        if (!$this->checkBelonging()) {
+            return "You don't own this room";
+        }
         $db=dbconnect();
-        $stmt2 = $db->prepare("DELETE FROM components WHERE id_room=?");
+        $stmt2 = $db->prepare("SELECT serial_number FROM components WHERE id_room = ?");
         $stmt2->bind_param("i", $this->id);
         $stmt2->execute();
+        $res = $stmt2->get_result();
+        if ($res->fetch_assoc()) {
+            return "You can't delete an empty room";
+        }
         $stmt2->close();
         $stmt1 = $db->prepare("DELETE FROM rooms WHERE id=?");
         $stmt1->bind_param("i", $this->id);
@@ -39,6 +47,10 @@ class Room
     }
 
     function rename($newName) {
+
+        if (!$this->checkBelonging()) {
+            return "You don't own this room";
+        }
 
         $db=dbconnect();
         $stmt = $db->prepare("UPDATE rooms SET name = ? WHERE id = ?");
@@ -57,6 +69,10 @@ class Room
      **/
     public function getAllFields(): array
     {
+        if (!$this->checkBelonging()) {
+            return "You don't own this room";
+        }
+
         $db = dbconnect();
         $stmt = $db->prepare("SELECT * FROM rooms WHERE id = ?");
         $stmt->bind_param("i", $this->id);
@@ -70,6 +86,11 @@ class Room
     }
 
     public function getComponents(){
+        
+        if (!$this->checkBelonging()) {
+            return "You don't own this room";
+        }
+        
         $db = dbconnect();
         $stmt = $db->prepare("SELECT id FROM components WHERE id_room = ?");
         $stmt->bind_param("i", $this->id);
@@ -86,6 +107,27 @@ class Room
     public function getId()
     {
         return $this->id;
+    }
+
+    public function checkBelonging() {
+        if (session_status() == PHP_SESSION_NONE) {
+            session_start();
+        }
+        
+        $db = dbconnect();
+        $stmt = $db->prepare("SELECT id_home FROM rooms WHERE id = ? LIMIT 1");
+        $stmt->bind_param("s", $this->id);
+        $stmt->execute();
+        $res = $stmt->get_result();
+        if ($row = $res->fetch_assoc()) {
+            if ($row['id_home'] != $_SESSION['home_id']) {
+                return false;
+            } else {
+                return true;
+            }
+        }
+
+        return false;
     }
 
 }
